pyproject.toml
[tool.poetry]
name = "trendmoon-mindshare-agent"
version = "0.1.0"
description = ""
authors = ["xiuxiuxar"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
open-aea-ledger-solana = ">=2.0.0,<3.0.0"
open-aea-ledger-cosmos = ">=2.0.0,<3.0.0"
open-aea-ledger-ethereum = ">=2.0.0,<3.0.0"
open-aea = ">=2.0.0,<3.0.0"
open-aea-cli-ipfs = ">=2.0.0,<3.0.0"
open-aea-test-autonomy = "==0.21.1"
open-autonomy = "==0.21.1"
click = "==8.1.8"
eth-account = "^0.13.5"
cowdao-cowpy = "1.0.1"
open-balpy = "0.1.6"
protobuf = ">=4.21.6,<4.25.0"
pydantic = "^2.11.3"
ecdsa = ">=0.15"
eth-abi = "==5.1.0"
eth-utils = ">=2.3.1"
eth-typing = "<=5.2.0"
hexbytes = "^1.3.1"
packaging = ">=23.1,<24.0"
pycryptodome = "==3.20.0"
requests = ">=2.28.1,<2.31.2"
web3 = ">=5.0.0"
aiohttp = ">=3.8.5,<4.0.0"
jsonschema = ">=4.3.3"
openapi-core = "==0.14.5"
openapi-spec-validator = "==0.2.8"
aioprometheus = ">=20.0.0,<21.0.0"
psutil = ">=5.9.5,<6.0.0"
dateparser = ">=1.1.1"
derive-client = "==0.2.29"
open-multicaller = ">=0.2.0,<0.4.0"
rich = "^14.1.0"
pandas-ta-classic = "^0.3.14b2"
httpx = "0.23.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]
